refactor: move tool call helpers into subdirectory
